src/main/java/com/platinum/MiServicio.java
package com.platinum;

import org.springframework.stereotype.Service;
import com.platinum.model.Ejecutivo;

@Service
public class MiServicio {
	public Ejecutivo obtenerInformacionEjecutivoPorRut(String rut) {
        // Implementación del método
        // ...
		
		
		Ejecutivo ejecutivo = obtenerEjecutivoPorRut(rut);
		
		if (ejecutivo == null) {
		}

        return ejecutivo;
    }

    // Este método hipotético obtiene un Ejecutivo por su rut.
    private Ejecutivo obtenerEjecutivoPorRut(String rut) {
        // Implementación de cómo obtener un ejecutivo por su rut.
        // ...

        return null; // Devuelve null por ahora, reemplázalo con la lógica real.
    }
}